+++ service/controller/v11/resource/deployment/resource.go
@@ -100,7 +100,7 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("deployment is in state '%s'", s))
 
 		if !key.IsSucceededProvisioningState(s) {
-			r.debugger.LogFailedDeployment(ctx, d, err)
+			r.debugger.LogFailedDeployment(ctx, d)
 		}
 		if !key.IsFinalProvisioningState(s) {
 			r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
@@ -123,15 +123,11 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 
 	res, err := deploymentsClient.CreateOrUpdate(ctx, key.ClusterID(customObject), mainDeploymentName, deployment)
 	if err != nil {
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("deployment failed; deployment: %#v", deployment), "stack", microerror.Stack(microerror.Mask(err)))
-
 		return microerror.Mask(err)
 	}
 
-	deploymentExtended, err := deploymentsClient.CreateOrUpdateResponder(res.Response())
+	_, err = deploymentsClient.CreateOrUpdateResponder(res.Response())
 	if err != nil {
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("deployment failed; deployment: %#v", deploymentExtended), "stack", microerror.Stack(microerror.Mask(err)))
-
 		return microerror.Mask(err)
 	}
