+++ service/controller/v12/debugger/debugger.go
@@ -2,6 +2,7 @@ package debugger
 
 import (
 	"context"
+	"io/ioutil"
 
 	"github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2018-05-01/resources"
 	"github.com/giantswarm/microerror"
@@ -30,16 +31,21 @@ func New(config Config) (*Debugger, error) {
 	return d, nil
 }
 
-func (d *Debugger) LogFailedDeployment(ctx context.Context, deployment resources.DeploymentExtended) {
+func (d *Debugger) LogFailedDeployment(ctx context.Context, deployment resources.DeploymentExtended, err error) {
 	if !key.IsFailedProvisioningState(*deployment.Properties.ProvisioningState) {
 		return
 	}
+
+	body, _ := ioutil.ReadAll(deployment.Body)
 
 	d.logger.LogCtx(ctx,
 		"correlationID", *deployment.Properties.CorrelationID,
 		"id", *deployment.ID,
 		"level", "error",
 		"message", "deployment failed",
+		"status", deployment.Status,
+		"body", string(body),
 		"name", *deployment.Name,
+		"stack", microerror.Stack(microerror.Mask(err)),
 	)
 }
